requirements.txt
numpy
torch
gymnasium[mujoco]
opencv-python
tensorboard
pyyaml
tqdm